pyproject.toml
[project]
name = "dhandho_bot"
version = "0.1.0"
description = ""
authors = [
    {name = "dilrong",email = "dilrong@dilrong.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "yfinance (>=0.2.54,<0.3.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-telegram-bot (>=21.11.1,<22.0.0)",
    "schedule (>=1.2.2,<2.0.0)",
    "lxml (>=5.3.1,<6.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
